Backend - fastapi/requirements.txt
fastapi
uvicorn
python-multipart
pymongo
python-dotenv
scikit-learn
nltk
python-docx
pytesseract
Pillow
sentence-transformers
numpy
PyMuPDF